Move Embedding generator base
