+++ python/samples/concepts/caching/semantic_caching.py
@@ -7,7 +7,7 @@ from dataclasses import dataclass, field
 from typing import Annotated
 from uuid import uuid4
 
-from semantic_kernel.connectors.ai.embeddings.embedding_generator_base import EmbeddingGeneratorBase
+from semantic_kernel.connectors.ai.embedding_generator_base import EmbeddingGeneratorBase
 from semantic_kernel.connectors.ai.open_ai.services.open_ai_chat_completion import OpenAIChatCompletion
 from semantic_kernel.connectors.ai.open_ai.services.open_ai_text_embedding import OpenAITextEmbedding
 from semantic_kernel.connectors.memory.in_memory.in_memory_store import InMemoryVectorStore
