+++ python/semantic_kernel/connectors/ai/google/google_ai/services/google_ai_text_embedding.py
@@ -8,7 +8,7 @@ from google.generativeai.types.text_types import BatchEmbeddingDict
 from numpy import array, ndarray
 from pydantic import ValidationError
 
-from semantic_kernel.connectors.ai.embeddings.embedding_generator_base import EmbeddingGeneratorBase
+from semantic_kernel.connectors.ai.embedding_generator_base import EmbeddingGeneratorBase
 from semantic_kernel.connectors.ai.google.google_ai.google_ai_prompt_execution_settings import (
     GoogleAIEmbeddingPromptExecutionSettings,
 )
