+++ python/tests/integration/embeddings/test_embedding_service.py
@@ -5,7 +5,7 @@ from typing import Any
 
 import pytest
 
-from semantic_kernel.connectors.ai.embeddings.embedding_generator_base import EmbeddingGeneratorBase
+from semantic_kernel.connectors.ai.embedding_generator_base import EmbeddingGeneratorBase
 from semantic_kernel.connectors.ai.prompt_execution_settings import PromptExecutionSettings
 from tests.integration.embeddings.test_embedding_service_base import (
     EmbeddingServiceTestBase,
